serial: imx: set CTS if DT property "linux, rs485-enabled-at-boot-time" is present
If we have a "linux,rs485-enabled-at-boot-time" DT property, we should
set CTS flag as soon as possible to the active state. This will set an
RS485 transmitter to receive-enable mode. Otherwise the RS485 bus will
be blocked for entire system boot time and other bus participants will
not be able to communicate.

Signed-off-by: Oleksij Rempel <o.rempel@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>