MIPS: loongson1: [OK] enable UART2 output
this version output a 'U' constantly from UART2

the UART2 is located as below, connect COM2 to
a PC, or attach U21's PIN.11 to a scope to see
the waveform.

---------------+
     ___       |
    -|       COM1
    -|         |
->Tx-| U21   COM2

at the same time, the LED8 and LED6 is blinking,
but maybe human not sensible, use a scope to "see"
it.

Signed-off-by: Du Huanpeng <u74147@gmail.com>